requirements.txt
jax
flax
optax
tensorflow_datasets
tensorflow_probability
ml_collections
wandb

